extracted auth panel option type
